+++ src/js/__tests__/index.js
@@ -8,6 +8,7 @@ import { JSDOM } from 'jsdom';
 const { window } = new JSDOM('<!doctype html><html><body></body></html>');
 global.window = window;
 global.document = window.document;
+global.HTMLElement = window.HTMLElement;
 global.navigator = global.window.navigator;
 
 global.window.matchMedia = () => ({